requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
boto3==1.35.10
botocore==1.35.10
Brotli==1.1.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.0
cssselect2==0.7.0
dj-database-url==2.2.0
Django==4.2.17
django-cors-headers==4.4.0
django-extensions==3.2.3
django-rest-knox==5.0.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
dynaconf==3.2.6
facebook-sdk==3.1.0
fonttools==4.55.3
future==1.0.0
google==3.0.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
httplib2==0.22.0
hvac==2.3.0
idna==3.8
inflection==0.5.1
Jinja2==3.1.4
jmespath==1.0.1
knox==0.1.14
kombu==5.4.1
loguru==0.7.2
MarkupSafe==2.1.5
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.1
paystackapi==2.1.3
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyparsing==3.1.4
pyphen==0.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-twitter==3.5
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
resend==2.4.0
rsa==4.9
s3transfer==0.10.2
shortuuid==1.0.13
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
twitter==1.19.6
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
validators==0.33.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
whitenoise==6.7.0
zopfli==0.2.3.post1